pyproject.toml
[project]
name = "blog-dashboard"
version = "0.1.0"
description = "Local blog dashboard for p.ocmatos.com"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "python-slugify>=8.0.0",
]

[project.scripts]
write = "dashboard.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dashboard"]
